+++ internal/errors/errors.go
@@ -14,12 +14,17 @@ import (
 func IsTimeoutError(err error) bool {
 	switch {
 	case
-		IsOpError(err, StatusTimeout),
-		IsOpError(err, StatusCancelled),
-		IsTransportError(err, TransportErrorCanceled),
-		IsTransportError(err, TransportErrorDeadlineExceeded),
-		errors.Is(err, context.DeadlineExceeded),
-		errors.Is(err, context.Canceled):
+		IsOpError(
+			err,
+			StatusTimeout,
+			StatusCancelled,
+		),
+		IsTransportError(err, TransportErrorCanceled, TransportErrorDeadlineExceeded),
+		Is(
+			err,
+			context.DeadlineExceeded,
+			context.Canceled,
+		):
 		return true
 	default:
 		return false
@@ -49,10 +54,18 @@ func As(err error, target interface{}) bool {
 	return errors.As(err, target)
 }
 
-// Is is a proxy to errors.Is
+// Is is a improved proxy to errors.Is
 // This need to single import errors
-func Is(err, target error) bool {
-	return errors.Is(err, target)
+func Is(err error, targets ...error) bool {
+	if len(targets) == 0 {
+		panic("empty targets")
+	}
+	for _, target := range targets {
+		if errors.Is(err, target) {
+			return true
+		}
+	}
+	return false
 }
 
 // New is a proxy to errors.New