+++ log/table.go
@@ -462,10 +462,10 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 		// nolint:govet
 		log := log.WithName(`pool`)
 		if details&trace.TablePoolLifeCycleEvents != 0 {
-			t.OnPoolInit = func(info trace.PoolInitStartInfo) func(trace.PoolInitDoneInfo) {
+			t.OnInit = func(info trace.TableInitStartInfo) func(trace.TableInitDoneInfo) {
 				log.Infof(`initialize start`)
 				start := time.Now()
-				return func(info trace.PoolInitDoneInfo) {
+				return func(info trace.TableInitDoneInfo) {
 					log.Infof(`initialize done {latency:"%v",size:{min:%d,max:%d}}`,
 						time.Since(start),
 						info.KeepAliveMinSize,
@@ -473,10 +473,10 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 					)
 				}
 			}
-			t.OnPoolClose = func(info trace.PoolCloseStartInfo) func(trace.PoolCloseDoneInfo) {
+			t.OnClose = func(info trace.TableCloseStartInfo) func(trace.TableCloseDoneInfo) {
 				log.Infof(`close start`)
 				start := time.Now()
-				return func(info trace.PoolCloseDoneInfo) {
+				return func(info trace.TableCloseDoneInfo) {
 					if info.Error == nil {
 						log.Infof(`close done {latency:"%v"}`,
 							time.Since(start),