+++ trace/scripting_gtrace.go
@@ -10,6 +10,30 @@ import (
 // both from t and x.
 func (t Scripting) Compose(x Scripting) (ret Scripting) {
 	switch {
+	case t.OnInit == nil:
+		ret.OnInit = x.OnInit
+	case x.OnInit == nil:
+		ret.OnInit = t.OnInit
+	default:
+		h1 := t.OnInit
+		h2 := x.OnInit
+		ret.OnInit = func(s ScriptingInitStartInfo) func(ScriptingInitDoneInfo) {
+			r1 := h1(s)
+			r2 := h2(s)
+			switch {
+			case r1 == nil:
+				return r2
+			case r2 == nil:
+				return r1
+			default:
+				return func(s ScriptingInitDoneInfo) {
+					r1(s)
+					r2(s)
+				}
+			}
+		}
+	}
+	switch {
 	case t.OnExecute == nil:
 		ret.OnExecute = x.OnExecute
 	case x.OnExecute == nil:
@@ -17,18 +41,18 @@ func (t Scripting) Compose(x Scripting) (ret Scripting) {
 	default:
 		h1 := t.OnExecute
 		h2 := x.OnExecute
-		ret.OnExecute = func(e ExecuteStartInfo) func(ExecuteDoneInfo) {
-			r1 := h1(e)
-			r2 := h2(e)
+		ret.OnExecute = func(s ScriptingExecuteStartInfo) func(ScriptingExecuteDoneInfo) {
+			r1 := h1(s)
+			r2 := h2(s)
 			switch {
 			case r1 == nil:
 				return r2
 			case r2 == nil:
 				return r1
 			default:
-				return func(e ExecuteDoneInfo) {
-					r1(e)
-					r2(e)
+				return func(s ScriptingExecuteDoneInfo) {
+					r1(s)
+					r2(s)
 				}
 			}
 		}
@@ -41,7 +65,7 @@ func (t Scripting) Compose(x Scripting) (ret Scripting) {
 	default:
 		h1 := t.OnStreamExecute
 		h2 := x.OnStreamExecute
-		ret.OnStreamExecute = func(s StreamExecuteStartInfo) func(StreamExecuteIntermediateInfo) func(StreamExecuteDoneInfo) {
+		ret.OnStreamExecute = func(s ScriptingStreamExecuteStartInfo) func(ScriptingStreamExecuteIntermediateInfo) func(ScriptingStreamExecuteDoneInfo) {
 			r1 := h1(s)
 			r2 := h2(s)
 			switch {
@@ -50,7 +74,7 @@ func (t Scripting) Compose(x Scripting) (ret Scripting) {
 			case r2 == nil:
 				return r1
 			default:
-				return func(s StreamExecuteIntermediateInfo) func(StreamExecuteDoneInfo) {
+				return func(s ScriptingStreamExecuteIntermediateInfo) func(ScriptingStreamExecuteDoneInfo) {
 					r11 := r1(s)
 					r21 := r2(s)
 					switch {
@@ -59,7 +83,7 @@ func (t Scripting) Compose(x Scripting) (ret Scripting) {
 					case r21 == nil:
 						return r11
 					default:
-						return func(s StreamExecuteDoneInfo) {
+						return func(s ScriptingStreamExecuteDoneInfo) {
 							r11(s)
 							r21(s)
 						}
@@ -76,18 +100,18 @@ func (t Scripting) Compose(x Scripting) (ret Scripting) {
 	default:
 		h1 := t.OnExplain
 		h2 := x.OnExplain
-		ret.OnExplain = func(e ExplainStartInfo) func(ExplainDoneInfo) {
-			r1 := h1(e)
-			r2 := h2(e)
+		ret.OnExplain = func(s ScriptingExplainStartInfo) func(ScriptingExplainDoneInfo) {
+			r1 := h1(s)
+			r2 := h2(s)
 			switch {
 			case r1 == nil:
 				return r2
 			case r2 == nil:
 				return r1
 			default:
-				return func(e ExplainDoneInfo) {
-					r1(e)
-					r2(e)
+				return func(s ScriptingExplainDoneInfo) {
+					r1(s)
+					r2(s)
 				}
 			}
 		}
@@ -118,58 +142,73 @@ func (t Scripting) Compose(x Scripting) (ret Scripting) {
 	}
 	return ret
 }
-func (t Scripting) onExecute(e ExecuteStartInfo) func(ExecuteDoneInfo) {
-	fn := t.OnExecute
+func (t Scripting) onInit(s ScriptingInitStartInfo) func(ScriptingInitDoneInfo) {
+	fn := t.OnInit
 	if fn == nil {
-		return func(ExecuteDoneInfo) {
+		return func(ScriptingInitDoneInfo) {
 			return
 		}
 	}
-	res := fn(e)
+	res := fn(s)
 	if res == nil {
-		return func(ExecuteDoneInfo) {
+		return func(ScriptingInitDoneInfo) {
 			return
 		}
 	}
 	return res
 }
-func (t Scripting) onStreamExecute(s StreamExecuteStartInfo) func(StreamExecuteIntermediateInfo) func(StreamExecuteDoneInfo) {
+func (t Scripting) onExecute(s ScriptingExecuteStartInfo) func(ScriptingExecuteDoneInfo) {
+	fn := t.OnExecute
+	if fn == nil {
+		return func(ScriptingExecuteDoneInfo) {
+			return
+		}
+	}
+	res := fn(s)
+	if res == nil {
+		return func(ScriptingExecuteDoneInfo) {
+			return
+		}
+	}
+	return res
+}
+func (t Scripting) onStreamExecute(s ScriptingStreamExecuteStartInfo) func(ScriptingStreamExecuteIntermediateInfo) func(ScriptingStreamExecuteDoneInfo) {
 	fn := t.OnStreamExecute
 	if fn == nil {
-		return func(StreamExecuteIntermediateInfo) func(StreamExecuteDoneInfo) {
-			return func(StreamExecuteDoneInfo) {
+		return func(ScriptingStreamExecuteIntermediateInfo) func(ScriptingStreamExecuteDoneInfo) {
+			return func(ScriptingStreamExecuteDoneInfo) {
 				return
 			}
 		}
 	}
 	res := fn(s)
 	if res == nil {
-		return func(StreamExecuteIntermediateInfo) func(StreamExecuteDoneInfo) {
-			return func(StreamExecuteDoneInfo) {
+		return func(ScriptingStreamExecuteIntermediateInfo) func(ScriptingStreamExecuteDoneInfo) {
+			return func(ScriptingStreamExecuteDoneInfo) {
 				return
 			}
 		}
 	}
-	return func(s StreamExecuteIntermediateInfo) func(StreamExecuteDoneInfo) {
+	return func(s ScriptingStreamExecuteIntermediateInfo) func(ScriptingStreamExecuteDoneInfo) {
 		res := res(s)
 		if res == nil {
-			return func(StreamExecuteDoneInfo) {
+			return func(ScriptingStreamExecuteDoneInfo) {
 				return
 			}
 		}
 		return res
 	}
 }
-func (t Scripting) onExplain(e ExplainStartInfo) func(ExplainDoneInfo) {
+func (t Scripting) onExplain(s ScriptingExplainStartInfo) func(ScriptingExplainDoneInfo) {
 	fn := t.OnExplain
 	if fn == nil {
-		return func(ExplainDoneInfo) {
+		return func(ScriptingExplainDoneInfo) {
 			return
 		}
 	}
-	res := fn(e)
+	res := fn(s)
 	if res == nil {
-		return func(ExplainDoneInfo) {
+		return func(ScriptingExplainDoneInfo) {
 			return
 		}
 	}
@@ -190,43 +229,53 @@ func (t Scripting) onClose(s ScriptingCloseStartInfo) func(ScriptingCloseDoneInf
 	}
 	return res
 }
+func ScriptingOnInit(t Scripting, c *context.Context) func(error) {
+	var p ScriptingInitStartInfo
+	p.Context = c
+	res := t.onInit(p)
+	return func(e error) {
+		var p ScriptingInitDoneInfo
+		p.Error = e
+		res(p)
+	}
+}
 func ScriptingOnExecute(t Scripting, c *context.Context, query string, parameters queryParameters) func(result result, _ error) {
-	var p ExecuteStartInfo
+	var p ScriptingExecuteStartInfo
 	p.Context = c
 	p.Query = query
 	p.Parameters = parameters
 	res := t.onExecute(p)
 	return func(result result, e error) {
-		var p ExecuteDoneInfo
+		var p ScriptingExecuteDoneInfo
 		p.Result = result
 		p.Error = e
 		res(p)
 	}
 }
 func ScriptingOnStreamExecute(t Scripting, c *context.Context, query string, parameters queryParameters) func(error) func(error) {
-	var p StreamExecuteStartInfo
+	var p ScriptingStreamExecuteStartInfo
 	p.Context = c
 	p.Query = query
 	p.Parameters = parameters
 	res := t.onStreamExecute(p)
 	return func(e error) func(error) {
-		var p StreamExecuteIntermediateInfo
+		var p ScriptingStreamExecuteIntermediateInfo
 		p.Error = e
 		res := res(p)
 		return func(e error) {
-			var p StreamExecuteDoneInfo
+			var p ScriptingStreamExecuteDoneInfo
 			p.Error = e
 			res(p)
 		}
 	}
 }
 func ScriptingOnExplain(t Scripting, c *context.Context, query string) func(plan string, _ error) {
-	var p ExplainStartInfo
+	var p ScriptingExplainStartInfo
 	p.Context = c
 	p.Query = query
 	res := t.onExplain(p)
 	return func(plan string, e error) {
-		var p ExplainDoneInfo
+		var p ScriptingExplainDoneInfo
 		p.Plan = plan
 		p.Error = e
 		res(p)